+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Effect: Ice Knife Collision Detection updates.
 - Switch to using DAE flag macros rather than itemacro if using newer versions of DAE.
 - Effect: Strahd Zombie Loathsome Limbs
+- Effect stub: Partially Amphibious.
 
 # 3.5.8